pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "neuroscript"
version = "2.0.0"
description = "Type-safe neural architecture composition system with capability-based shape inference and static validation"
readme = "README.md"
authors = [
    {name = "NeuroScript Contributors"}
]
license = {text = "Apache License 2.0"}
requires-python = ">=3.8"
keywords = ["neural", "architecture", "validation", "shape-inference", "composition"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pyyaml>=6.0",
    "jinja2>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=22.0",
    "mypy>=0.990",
    "ruff>=0.1.0",
]
full = [
    "torch>=2.0",
    "psutil>=5.9",
]
cuda = [
    "torch>=2.0",
]

[project.urls]
Homepage = "https://github.com/anthropics/neuroscript"
Documentation = "https://github.com/anthropics/neuroscript/blob/main/README.md"
Repository = "https://github.com/anthropics/neuroscript.git"
"Issue Tracker" = "https://github.com/anthropics/neuroscript/issues"

[tool.setuptools]
packages = ["core", "tools"]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
src = ["py.typed"]

[tool.black]
line-length = 100
target-version = ['py38']
include = '\.pyi?$'

[tool.pytest.ini_options]
testpaths = ["test"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers --tb=short"
markers = [
    "unit: unit tests",
    "integration: integration tests",
]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false

[tool.ruff]
select = ["E", "F", "W", "I"]
ignore = ["E501"]  # Line too long (handled by black)
line-length = 100
